@@ -1,19 +1,16 @@
 {
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build"
-  },
-  "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "^2.0.2",
-    "@tsconfig/svelte": "^3.0.0",
-    "fontawesome-svgs": "^1.0.2",
-    "svelte": "^3.55.1",
-    "svelte-check": "^2.10.3",
-    "tslib": "^2.5.0",
-    "typescript": "^4.9.5",
-    "helion": "^1.0.8",
-    "vite": "^4.1.1"
-  }
+	"private": true,
+	"type": "module",
+	"scripts": {
+		"dev": "vite",
+		"build": "vite build"
+	},
+	"devDependencies": {
+		"@sveltejs/vite-plugin-svelte": "^6.2.4",
+		"fontawesome-svgs": "^1.0.2",
+		"svelte": "^5.50.0",
+		"vite": "^7.3.1",
+		"@tailwindcss/vite": "^4.1.18",
+		"tailwindcss": "^4.1.18"
+	}
 }
