@@ -1,11 +1,15 @@
-<h1>Pixel Text Generator</h1>
-<h2>Links</h2>
-<ul>
-	<li><a href="https://github.com/TheCymaera/pixel-text-generator" target="_blank">Source code on GitHub</a></li>
-</ul>
+<div class="typography">
+	<h1>Pixel Text Generator</h1>
+	<h2>Links</h2>
+	<ul>
+		<li>
+			<a href="https://github.com/TheCymaera/pixel-text-generator" target="_blank">Source code on GitHub</a>
+		</li>
+	</ul>
 
-<h2>License</h2>
-<p>
-	All content generated using the tool can be used for commercial or non-commercial purposes. <wbr>
-	Attribution is appreciated but not due.
-</p>
+	<h2>License</h2>
+	<p>
+		All content generated using the tool can be used for commercial or non-commercial purposes. <wbr>
+		Attribution is appreciated but not due.
+	</p>
+</div>
