@@ -1,17 +1,6 @@
-import {} from "helion/core.js";
-import {} from "helion/LightTheme.js";
-import {} from "helion/AspectRatio.js";
-import {} from "helion/Stack.js";
-import {} from "helion/Intangible.js";
-import {} from "helion/AppBar.js";
-import {} from "helion/Panel.js";
-import {} from "helion/CircleButton.js";
-import {} from "helion/OutlinedTextField.js";
-import {} from "helion/ColorPickerField.js";
-import {} from "helion/StandardView.js";
-
 import App from "./App.svelte";
+import { mount } from "svelte";
+import "./main.css";
+import "./typography.css";
 
-new App({
-	target: document.body,
-});
+mount(App, { target: document.body });
