@@ -1,14 +1,11 @@
 import { defineConfig } from 'vite';
 import { svelte } from '@sveltejs/vite-plugin-svelte';
+import tailwindcss from '@tailwindcss/vite';
 
 export default defineConfig({
 	base: "./",
-	build: {
-		modulePreload: {
-			polyfill: false,
-		}
-	},
 	plugins: [
 		svelte(),
+		tailwindcss(),
 	],
 });
